requirements.txt
opencv-python-headless
ultralytics
paho-mqtt
pandas
altair
torch
torchvision
